+++ src/signals/incident-management/containers/IncidentOverviewPage/components/Filter/style.scss
