+++ src/signals/incident-management/containers/IncidentOverviewPage/components/List/style.scss
