+++ src/signals/incident-management/containers/IncidentOverviewPage/components/Pager/style.scss
