+++ src/signals/incident-management/containers/IncidentOverviewPage/reducer.js
