+++ src/signals/incident-management/containers/IncidentOverviewPage/style.scss
